Move Spinner component out of micro-cluster
