+++ app/javascript/src/admin/components/Spinner.jsx
